knowledge_base_api/pyproject.toml
[tool.poetry]
name = "knowledge-base-api"
version = "4.0.0"
description = "A production-ready FastAPI API for a multi-tenant knowledge base."
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
# Указываем, что проект совместим с Python 3.9 и выше, но не с версией 4.0
python = ">=3.9,<4.0"

# Основные зависимости проекта
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.27"}
asyncpg = "^0.29.0"
pydantic = "^2.6.3"
pydantic-settings = "^2.2.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
argon2-cffi = "^23.1.0"
aioboto3 = "^12.3.0"
python-json-logger = "^2.0.7"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
# Зависимости, необходимые только для разработки и тестирования
pytest = "^8.0.2"
pytest-asyncio = "^0.23.5"
httpx = "^0.27.0"
ruff = "^0.2.2"
alembic = "^1.13.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Конфигурация для линтера и форматтера Ruff
line-length = 88
indent-width = 4

[tool.ruff.lint]
# Включаем правила для проверки стиля (аналог flake8) и сортировки импортов (isort)
select = ["E", "F", "W", "I"]
ignore = []

# Директории, которые не нужно проверять
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]